media: cros-codecs: vaapi: store pending VA picture into GenericBackendHandle
Pending VA pictures were stored in a pending job queue, whereas the
GenericBackendHandle only had their surface ID and received the picture
itself only when it was completed.

This design adds some dependencies between the handles and the backend,
and it makes more sense for the GenericBackendHandle to own the VA
pictures from the beginning, and transition pending pictures to the
ready state itself. The progression of PictureState from pending to
ready is made more obvious that way, and the state management can now be
centralized in GenericBackendHandle, removing duplicate code.

This also allows us to remove the PendingJob struct and replace it by
counted references to the underlying GenericBackendHandle.

BUG=b:214478588
TEST=cargo test --features vaapi -p cros-codecs

Change-Id: Id8b05ccd1524f603c61231d368a9fd234247840e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4164544
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Daniel Almeida <daniel.almeida@collabora.corp-partner.google.com>