decoder/h264: forget buffers before the first sps
We were storing the buffers coming before the first sps and submitted
them for decoding after the first sps arrived: this results in the
buffers being processed out-of-order and will likely corrupt the
sequence.

Just drop the buffers until we know the parameters for the sequence,
just like vp8 and vp9 do ; this is the correct way to do and it also
makes the 3 codecs closer.