decoders/h264: make the decoder manage the current picture
The current implementation delegates the current picture management to
the backend, which results in duplicate management of a potentially
dangerous Option type. Move it to the decoder where it can be
centralized instead, and pass the current picture by reference to the
backend when needed.

This also has the advantage of removing the current_picture member from
the vaapi backend structure and making all the vaapi backends identical,
opening the way for further simplification.