decoders/vaapi/tests: move VA struct checking to their own tests
These checks are very intrusive to the backend and decoder since they
have a side-effect and effectively steal the decoded images, preventing
the decoder from behaving as it would in real conditions.

Move these into a dedicated test, and also limit them to the test-25fps
stream since we are just checking that the construction of a structure
is done correctly.