decoders: fix display order of frames in non-blocking mode
Making the tests check the order of frames revealed that frames were
sometimes presented out-of-order when decoding in non-blocking mode,
i.e. when there can be more than one frame in the ready_queue of the
decoder.

The cause is that we test all the frames in the queue for readiness, and
return all those that pass the test. This leaves space for a race
condition where frame X is not ready at the time of testing, and frame
X + 1 becomes ready by the time it is tested. When this happens, frame
X + 1 is returned before frame X.

Fix this by stopping filtering on the first frame that signals it is not
ready. Since the decoder processes the jobs we give it in-order, this
guarantees that the frames will always be returned in the correct order.