media: cros-codecs: simplify the signature of Nalu::next()
Currently, Nalu::next() takes two arguments, a Cursor<T> (to keep track
of the current position in the data), and the data itself (T) so that it
can be assigned to Nalu<T> for further use.

This signature is confusing, not only because it is repetitive, but most
importantly, because it allows Cursor<T> and T to be unrelated, which is
not the intended usage. To fix this, extract T from the cursor itself by
cloning it and calling into_inner().

This does not add any new costs, as presumably the call to clone() is
simply moved into the method directly.

Crosvm uses &[u8] directly, but other users are free to use other types,
such as the ones offered by the Bytes crate to do away with lifetime
issues in an efficient (zero-copy) manner. In such case, clone() remains
cheap.

BUG=b:214478588 TEST="cd media/cros-codecs && cargo test --release
--features vaapi -- --include-ignored"

Change-Id: Id827c39c507d8bc8766269e8001c4895b7d3f98d