decoder/h264: simplify ready queue
The DPB ensures that frames are added into the ready queue in display
order, so we don't need to have a sorting mechanism of our own.

Reordering the ready queue can actually be dangerous if e.g. we start a
new sequence (resetting the pic_order_cnt to 0) before consuming the
frames from the previous sequence: in this case the new latest frame
will come out first, whereas this problem does not exist with a simple
FIFO.