decoders: remove poll() method
Now that decoded handles can be synced and queried for readyness, we
don't need a method that potentially waits on all the decoded frames to
be ready - we can just grab as many handles as we need and wait for
them.

Removing this method allows us to drop the pending_jobs queue of the
VAAPI backend, which was basically a mirror of the ready_queue of
decoders and had the potential to fall out of sync and make us fall into
an invalid state.