stop importing anyhow::Result
We don't want to encourage the use of anyhow::Result for all functions ;
actually most should return their own error type to allow for precise
matching and error reporting.

Use of anyhow should be the exception rather than the rule, so stop
importing it and request all users to mention it explicitly. This makes
it easier to understand what a function returns and is not too long
anyway.