decoders: update stateless decoding API
The current decoding API was creating a (non-mandatory) dependency
between the queuing of decoding work and the dequeuing of decoded
frames. The fact that is tried its best not to fail also resulted in
some complex internal state, and despite that the decoders were
sometimes running out of resources.

Change this to an API where the submission of decoding jobs and the
retrieval of decoded frames are completely separate. decode() is now
allowed to return an error and to request the caller to address the
issue before re-calling it with the same data, which removes the
internal state complexity.

decode() can now fail for two reasons:

1) There are no available output resource to submit a job to the
   hardware, or
2) The stream format has changed.

In both cases, the client is responsible for processing all pending
events (either decoded frames or stream format change notifications)
before calling decode() again with the same data, which should this time
result in a success.

This API is well-suited to stateless hardware, and we can easily build a
stateful API on top of it.

More importantly, it reduces the complexity of our encoders and
ensures that all resource management issues can be addressed by the
client.