decoder: move format management burden from backend to decoder
A backend should have as little state as possible - notably, it should
not worry about whether a format can be set at a given time, since the
decoder is already tracking that state. Change the format setting
methods to take the expected format information provider as a parameter
instead of storing it in the backend itself. This way the decoder can
only call these methods if it already has that information, and we can
remove state from the backend that is redundant with that of the
decoder.

Format setting is not handled yet, but this will make it easier to
implement.