utils: fix h264 frame splitter
The parsing of first_mb_in_slice was incorrect because we called
as_ref() on the data member of AccessUnitParser, which does not take the
slice offset into account as the Nalu as_ref() method does. Use the
latter in order to get the correct first_mb_in_slice value. This
prevents all of Fluster's H.264 tests from failing.