vaapi: do not set the surface fourcc attribute
Setting the surface's fourcc attribute changes the internal format the
frames are rendered into. This can cause several issues:

* When creating an image, if the same fourcc is applied when the pixels
  are mangled a second time on AMD, resulting in an incorrect frame,

* On Intel, setting the fourcc to I420 results in vaEndPicture returning
  an INVALID_PARAMETER error.

Let's not specify this parameter for now - libva will pick whichever
format is best, and we will get the right ordering anyway when we create
the vaImage (as long as we don't derive it).

This results in proper decoding with NV12 and I420, on both Intel and
AMD.