ccdec: do not assume h264 as the default coded format
H.264 is very popular, but it is still not a good idea to simply assume
it as a default.

Passing some random data as an h.264 stream crashes cros-codecs, or
produces no useful output as no valid NALUs are parsed. This is the
best-case scenario, but we should not rely on that.

Leave the responsibility of properly labelling the data to the user.