decoders: vp9: enable inter-frame DRC
This is supported by the VP9 codec. Enable it at the decoder level so
that the backend can proceed accordingly, if it supports it.

Backends can signal whether this is supported because they can access
the frame data on new_sequence. This lets them determine whether it is
a non-key-frame drc and reject it if needed.

Notice that the negotiation code had to be rewritten with super-frames
in mind. With SVC, the first frame will usually be a key-frame, with
inter-frames carrying the other layers.

We do not want each of those to be considered as a separate DRC event.
Not only that, allowing them to be will cause an infinite loop.

Instead, negotiate based on the largest spatial layer. That will be
enough to house the other layers in between.