utils: vaapi: reinstate coded_resolution on ParsedStreamMetadata
This field was removed in favor of the coded_resolution on the surface
pool, as they used to be the same.

Now, the pool can be larger than the current coded resolution for a
particular frame, as a larger surface can always house a smaller
picture, saving on reallocations. This is particularly helpful for
SVC streams.

Fix this by reinstating the field such that we can always know the
coded resolution in use when a particular frame was decoded.