decoders: h264: fix frame_num value for gaps
The entire point of the decoding process for gaps in frame_num is to
compute and assign missing frame_num values for "nonexisting" pictures.

The same value of frame_num was being used for all of them. This would
crash the decoder as the stream would later try to retrieve the pictures
from the dpb using the correct number.

Fix this by using the number computed by equation 7-24 instead.

See 8.2.5.2 in the specification for more details.

Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>