decoders: stateless: standardize flush() across codecs.
If we take GStreamer as an inspiration, then flush() should return
whatever is ready while resetting the decoder's DPB.

This is what we had implemented in h.264 already, so merely standardize
the above by enacting the same semantics in the other codecs.

After flushing, the decoder will need a new key frame to continue.