examples: ccdec: optionally output to multiple files
This is *very* handy to visualize and/or debug DRC streams. The reason
being that there is usually no way to instruct visualization software
(like YUView or RawPixels) to switch resolutions on the fly with a
single file.

The strategy used here is similar to the multifilesink gstreamer
element. This allows users to input the resolution on a frame-by-frame
basis when loading the YUV data into other programs.