vaapi: move surface allocation to the surface pool
This will be useful for decoding into user-provided memory, as we will
be able to pass a reference to the pool and let the user take care of
how many buffers shall be allocated, when, and in what format. The
current way of doing is suitable for self-allocated buffers, but not so
much for importing.