decoder: remove format getting methods
The format was set to the mapped format of VAImage, which is not useful
for the case where we want to import external memory and need to
calculate the required space, as for this we need to know the internal
format of the decoder, which is going to be a (fourcc, modifier) tuple.

These methods are not used in the current code, so just remove them for
now so they don't get in the way of new code - we will add the
equivalents once we figure out how format should be managed.