decoder/h264: move static methods out of generic type impl
These methods do not need to be inside the type, and having them there
forces us to specify the generic parameters every time (and maybe even
generates duplicate code?).

Move them to the module level and keep them private - as they are
supposed to only be used by sub-modules, this is actually the correct place
for them.