decoders: h264: fix scaling list zigzag scan
The algorithm for scanning the scaling lists was wrong. The zigzag
matrices dictate the placement in the destination array, but the
original code scanned the source array in zigzag order instead.

This is now the same algorithm as implemented in Chromium.

Improves the score for the JVT-FR-EXT test suite from 12 to 42/69.