vaapi: introduce pooled surface type
The way surfaces were returned to their pool relied on the good will of
the borrower, and had potential for user mistakes. Since we want to keep
better accounting of our surfaces, take advantage of the fact that
libva's Picture type can accept any type that contains a Surface to move
the borrowing/return semantics to a PooledSurface type.

This allows to make sure that the client does not leak surfaces
unexpectedly, and will allow us to track exactly how many surfaces a
given pool is managing.