vaapi: add pool surfaces accounting
In order to decide a surface allocation strategy, the client needs to
know how many total surfaces the pool is currently managing, not just how
many surfaces are free.

Keep accounting of all our surfaces, borrowed or not, by indexing them
in a hash map using their VA ID. We also keep the resolution of each
surface to we can remove them if needed when the coded resolution
changes, and not when the borrowed surface is eventually dropped, which
can lead to race conditions if e.g. the resolution went down and then up
again, in which case we would end up with more surfaces than we wanted.