decoder: introduce StreamInfo struct
Several properties of the current stream were made available as separate
methods. Group them into a single stream_info method that returns a
structure of the same name with all these parameters. This highlights
the fact that they are static properties of the current stream, and not
dynamic parameters of the decoder that can be changed.