decoder/stateless: expose the SurfacePool to the client
Expose the surface pool to the client, in particular when a resolution
change event occurs. This will let the client decide how many target
buffers should be allocated and when to allocate them, something that
the previous design did not allow. More importantly, it will also allow
the client to provide its own memory backings for surfaces.