lib: add support for user-allocated memory
Add support for decoding into user-allocated memory. This memory type is
not recommended in practice, as e.g. VAAPI will not decode properly into
it, but it is added for the sake of exhaustiveness.

Rendering into user memory requires for now the use of a dedicated
struct that allocates the memory while guaranteeing that the
requirements of the underlying backends (VAAPI for now) in terms of
alignment and size are respected.