decoder: make DecodedHandle generic over memory descriptor
Make DecodedHandle generic over the memory descriptor and make is able
to retrieve the descriptor passed when creating a buffer.

This is important for scenarios involving external memory as without it
the user cannot know which attached descriptor corresponds to a decoded
frame.

As a result of this, the dummy backend is restricted to implementing
DecodedHandle<()>. This is a needed limitation since it doesn't store
the descriptor, and is perfectly fine since we only use it with the ()
memory type anyway.