decoder: allow memory mappings to be created from non-mutable handles
Memory mappings for decoded frames are read-only. The only reason why
creating one required a mutable reference was because we called the
sync() method, that does a mutable borrow,  in order to ensure the
memory was up-to-date.

Move the burden or calling sync() to the caller, and return an error if
this has not be done instead. This allows the caller to access the
resource and the mapping at the same time if needed, something that
would not be possible is dyn_picture() did a mutable borrow.