decoder/h264: call submit_picture from finish_picture
These methods are always called one after the other and are closely
related, so call submit_picture from finish_picture instead of expecting
them to be called sequentially.

Doing so is slightly safer as it also forces the caller to take and
check the current picture instead of doing an unwrap in submit_picture.

Fixes #31 (albeit not ideal, but better fix to follow after more
reworking).