decoder: add method to hold surfaces from the pool
The SurfacePool assumes that all its surfaces are available to be
decoded into, but the client might have different needs, e.g. it might
need to wait until a given surface is explicitly marked as ready to
allow it to be decoded into.

Add a method to SurfacePool that allows us to take a surface out of the
pool temporarily in order to allow clients to implement this kind of
workflow.