decoder: make the memory descriptor parameter of DecodedHandle an attribute type
Types should implement DecodedHandle only once. This allows us to move
the memory descriptor parameter of DecodedHandle into an attribute.

Doing so brings the benefit of being able to access the descriptor type
from other types, like StatelessDecoderBackend, which doesn't need to
have this generic parameter repeated. It is also a requirement to design
a single stateless decoder structure that can be generic against the
codec type and backend.