decoder/stateless: add generic decoder structure
A number of fields in the stateless decoders are duplicated between
codecs and can be factorized. This patch introduces a generic stateless
decoder structure that is generic against a codec and backend type,
allowing to create decoders specialized to a given codec and backend as
follows:

    let decoder: StatelessDecoder<H264, VaapiBackend<()>>;

The following patches will convert some of the decoders to this new
type.