decoder: make decode() return processed length
Require decode() to return how much of the input buffer has actually
been consumed. This will be useful to signal the caller that the input
was not correctly split and to simplify our decoding loop.

For now, just return the size of the input buffer since that's what we
are effectively consuming.