decoder/vp9: fix returned value of decode()
Always return the size of the passed input instead of trying to compute
it ourselves. The VP9 specification requires that the last byte of the
input is used to determine whether the chunk contains a superframe or
not, so if parsing is succcessful then the processed input will always
be the exact same size as the passed input.