decoders: h264: possibly renegotiate on every picture
A picture may refer to a SPS that is not the current one in use. Thus,
change the renegotiation logic to check whether it is possible to
renegotiate on every picture.

If the decoder is in the AwaitingStreamInfo state, we renegotiate with
the first SPS seen so that we can at least initialize the backend and
transition into the AwaitingFormat state.

If a picture references a SPS other than the first one parsed, a
renegotiation will happen in begin_picture() when the actual sps for the
picture is chosen.

Lastly, notice that a header can theoretically refer to another SPS and
PPS than the ones currently in use. If so, change them, but make sure we
do not renegotiate mid-picture.