decoders: h264: try to parse PPS again on error
If a PPS comes in but fails to parse, try to parse it again when
a new SPS comes along. This condition was highlighted by one of the
h.265 tests and this code pattern was used to fix the issue there.

This allows us to work around broken streams in which the PPS is sent
before the SPS for whatever reason.