decoders: h264: check that first_mb_in_slice increases monotonically
This checks a assumption in the specification for slices pertaining
to the same picture:

– If separate_colour_plane_flag is equal to 0, the value of
first_mb_in_slice shall not be less than the value of first_mb_in_slice
for any other slice of the current picture that precedes the current
slice in decoding order.

– Otherwise (separate_colour_plane_flag is equal to 1), the value of
first_mb_in_slice shall not be less than the value of first_mb_in_slice
for any other slice of the current picture that precedes the current
slice in decoding order and has the same value of colour_plane_id.