decoder/h264: remove cur_pps_id
The PPS of the current picture is stored in its state, so we can just
refer to it when needed. The only other context when we need to access
the PPS, the frame_num_gap handling method, can receive it by argument
from its caller, which allows us to get rid of an extra lookup.