decoder/stateless: move Picture associated type to its own trait
Backends may need to store codec-specific data on a per-frame basis. So
far this was achieved in the VAAPI backend through a generic parameter,
but this complicates things and introduces state at the stream level
while the state should not survive the frame boundary.

Prepare the ground for improving this by moving the Picture associated
type out of the StatelessDecoderBackend trait into its own
StatelessDecoderBackendPicture trait. This allows backends to keep one
implementation of StatelessDecoderBackend for all codecs, while allowing
them to specify a Picture type on a per-codec basis, which is exactly
the granularity we need for codec-specific data.