decoders: add support for multiple pools
We need to support new pools to cater to SVC content where multiple spatial
layers are being decoded, otherwise we would have to clear bigger frames before
decoding smaller layers to avoid leaving trash in them and breaking the decode
process.

This patch sets the stage by allowing multiple pools to exists, but only a
single pool is created in the VA code for now. Subsequent patches will build on
this one to actually allocate multiple pools depending on the content being
decoded.