vaapi: av1: submit slice parameters as an array
All tiles belonging to the same tile group must be submitted to the vaapi
driver as an array.

The Intel driver, in particular, expects one array of Slice Parameters per
instance of Slice Data when submitting the data for an AV1 Tile Group. It will
produce a corrupted frame if the slice parameters are submitted separately
(i.e.: in their own VABuffers).

This was discovered by Nicolas Dufresne.