backend: vaapi: add a handle for post-processing
Add a new handle for post-processing. This use-case requires a surface for
non-filtered data (for the decode process), and one surface for the filtered
data (for display).

This was previously impossible, as we only had one handle. Thus add a
new type of handle and unite them through an enum.

A next commit will make use of this handle to implement film grain in the AV1
VA-API code.