decoders: vaapi: av1: add film grain support
Use the new VAAPI handle to add film grain support in the VAAPI AV1 decoder.

One surface contains the grain and is meant to be displayed, while the other one
does not. We need the decode surface to not have any grain, otherwise this would
break the decoding process.

The number of surfaces is doubled to accomodate for this, but only when film
grain is active.