vaapi: move Rc'd part of the SurfacePool into a hidden inner member
This simplifies the types a bit as a SurfacePool does not always need to
be wrapped into a Rc<RefCell<>> and the client does not need to
constantly borrow from it. Furthermore this also simplifies the
get_surface() method, which doesn't require the caller to pass the same
pool as `self` as argument.