decoder/h264/vaapi: do not sync first frame when decoding interlaced stream
Since VAAPI uses fences on its surfaces, it will wait for all pending
operations to be finished before initiating a new on that depends on the
first surface, so this explicit sync should not be needed.

Tested by passing the whole H.264 Fluster suites in non-blocking mode,
on both Intel and AMD, without any regression.

Doing so should allow us to simplify the design of the VAAPI handle a
bit, as it won't be able to be in sync state during decoding.