codec/h264: simplify bumping process
Frames can be removed from the DPB when they are not a reference and are
not needed for output. Currently we check that condition in
`remove_unused` as well as in `bump`, but we can simplify things by just
calling `remove_unused` whenever we have bumped and need to free some
space.

This also allows us to remove the `flush` argument from `bump`, since
the only caller with this set to `true` can just call `clear` by itself
to achieve the same result with better efficiency.