codec/h264: separate DPB entry reference and frame
The DPB serves two purposes: as a way to reorder frames in display
order, and as a way to mark frames that are still used as reference.

Currently both purposes are served by the same `handle` member, but
separating it into dedicated members brings benefits: the type of the
decoded frame could be different from the one used for reference, and
the decoded frame could be take()n away from an Option when the frame is
bumped, making sure it cannot be emitted a second time.

IOW, this should allow the removal of the Rcs to reference frames in the
DPB, once a proper reference frame type is introduced.