codec/h264: merge is_second_field and other_field into single type
Having two members to describe the field rank is confusing and carries
the risk that both won't be updated or consulted at the same time. Merge
them into a single member that can be trusted to reliably carry the
state of the field as well as the reference to the other one, if any.