codec/h264: make sure field information is correct when passing PictureData to the DPB
The DPB used to work with Rc<RefCell<PictureData>>, but the problem with
this type is that it can be arbitrarily created while the PictureData is
incomplete - namely, when the `field_rank` of the first field is not
set, which can only be set after the second field is fully parsed.

Introduce a new type that can only be created after the field
information is correctly set by a conversion method. This makes all
field management transparent and guarantees that the DPB is working with
correct data.