decoder/stateless/h264: remove last_field
This field is only used with interlaced streams, and in case the DPB is
full when we try to add the first field of a frame to it.

It has been confirmed with Daniel Almeida (who initially wrote this
code) that this field was added as an extra safety to mimic something
GStreamer does, but is not warranted by the specification itself.

In practice, none of our unit tests or Fluster streams ever exercise a
path where last_field is set, and its presence seriously complicates the
code. Let's remove it.