codec/h264/dpb: filter long term refs using iterator
The explicit call to retain was a bit ambiguous if one didn't notice
that the short term refs list was already filtered, as it could imply
that elements before num_short_term_refs were removed and thus
invalidate its value.

Improve this by filtering entries that are second fields from the long
term refs iterator, just like we do for the short term refs one. This is
both clearer and more consistent.