encoder/h264: Use deduplicated version of LowDelay predictor
This change removes duplicated code from H264's LowDelay predictor. With
this change LowDelay tail size greater then 1 is no longer supported.
This may be added in the future, however the contribution to compression
performance may not be worth it.

With this commit the duplication process of LowDelay predictor is done.