decoder/stateless: add a poll FD that signals whenever an event is pending
Add a `poll_fd` method to `StatelessVideoDecoder` returning a file
descriptor that signals whenever an event is pending.

The descriptor is itself implemented by merging two `EventFd`s: one on
the ready frames queue (for `FrameReady` events), the other one on the
decoder state (for `FormatChanged` events).

That way, a client can poll for decoder events instead of busy-waiting
on them.

This change also introduces potential errors to constructors that never
returned any, so the decoder builders are updated accordingly.