decoder/stateless: specialize StatelessVideoDecoder via its handle and frame pool type
We were taking the whole backend as a generic argument, only to access
these two types. This is incorrect since `StatelessDecoderBackend` is
only used in the context of `StatelessDecoder`, which is just one of the
possible instances of `StatelessVideoDecoder`.

Specializing against the actually required types will allow us to create
trait objects from the `StatelessVideoDecoder` interface, which will
prove useful for user programs.