decoder: enable fully dynamic StatelessVideoDecoder trait objects
Add the necessary implementations and wrapper types to allow a
StatelessVideoDecoder to be turned into a fully dynamic trait object
with no other dependency that its Descriptor.

This now allows user code to control any instance of a decoder,
regardless of its codec or backend, through the same code path.