codec/h264/parser: make pic_height_in_map_units_minus1 a u16
The H.264 specification mentions that the maximum height of a stream is
4320 pixels (level 6.2). This means that pic_height_in_map_units_minus1
cannot exceed 270, so change its type to u16 and reject input values
that won't fit.

This has the benefit of bounding the possible values for
pic_height_in_map_units_minus1 in such a way that overflows in future
operations become impossible.