codec/h264/parser: make pic_height_in_map_units_minus1 a u8
The H.264 specification mentions that the maximum height of a stream is
2304 pixels. This means that pic_height_in_map_units_minus1 cannot
exceed 255, so change its type to u8 and reject input values that won't
fit.

This has the benefit of bounding the possible values for
pic_height_in_map_units_minus1 in such a way that overflows in future
operations become impossible.