codec/h264/parser: make pic_width_in_mbs_minus1 a u8
The H.264 specification mentions that the maximum width of a stream is
4096 pixels. This means that pic_width_in_mbs_minus1 cannot exceed 255,
so change its type to u8 and reject input values that won't fit.

This has the benefit of bounding the possible values for
pic_width_in_mbs_minus1 in such a way that overflows in future
operations become impossible.