codec/h264/parser: make pic_width_in_mbs_minus1 a u16
The H.264 specification mentions that the maximum width of a stream is
8192 pixels (level 6.2). This means that pic_width_in_mbs_minus1 cannot
exceed (8192 / 16) - 1, so change its type to u16 and reject input
values that won't fit.

This has the benefit of bounding the possible values for
pic_width_in_mbs_minus1 in such a way that overflows in future
operations become impossible.