decoder/stateless/vp8: let the backend decide which resources it needs
The decoder code was checking explicitly for output buffers
availability before decoding frames, but that requirement is not needed
by all backends - for example V4L2 stateless has different requirements.

Move this check to the backend by introducing a new_picture() method
that is responsible for allocating the resources required for a picture
and returning the expected error if it cannot secure them.