decoder/stateless/av1: process one OBU per decode call
decode() was trying to process a whole temporal unit, which complicates
the code quite a bit: the decoder must notably check for available
resources for all the frames in the OBU, which requires to parse the
input several times.

Our contract is that the caller must check how much input has been
consumed and re-submit the remainder if it has not been entirely
processed. Leverage this and only process one OBU per call to decode(),
which simplifies the resource availability check and will also allow us
to move it into the backend.