decoder/stateless: add utility method to wait for the next event
The proper way to wait for the next event is the poll the decoder's
event file descriptor instead of calling decode() over and over again
until an event is delivered, like our simple_decode_loop() helper
currently does.

Provide a decoder helper that does that and make it available as the
public interface. While probably not useful for real-life scenarios, it
is handy to have.

This also allows us to timeout instead of looping infinitely if the
decoder is starved for output buffers, as currently happens with
Fluster's RPS_E_qualcomm_5 H.265 test vector.