Fix AV1 frame size override logic
When the frame size override flag is set to true, the parser should read
the `found_ref` bits to find a reference frame to set the frame size to.
Currently this bit is interpretted to correspond directly with a
reference frame, but it should actually correspond to an entry in the
`ref_frame_idx` table, which in turn corresponds to a reference frame.

Bug: 382695401
Test: video.PlatformDecoding.cros_codecs_vaapi_av1_files_from_bugs_346405213
Change-Id: Id7a0db51d5ca879d2e84674678248721807472fa