Add AV1 frame size override support
Previously the AV1 stateless decoder made no distinction between visible
resolution and coded resolution, and would only change the coded
resolution when it received a new SequenceHeaderObu. This would not
allow for DRC through AV1's frame_size_override flag.

This change makes it so that the AV1 stateless decoder keeps track of
the visible resolution and will trigger a resolution change event if the
incoming frame resolution does not match the previous visible
resolution.

Bug: 382695401
Test: video.PlatformDecoding.cros_codecs_vaapi_av1_files_from_bugs_346405213
Change-Id: I6cc2ee982da46b53ed74cc963ad39e0f9389a571