Rewrite V4L2 ccdec using C2Wrapper
Note that this CL may trigger an apparent regression in H264 cros-codecs
tests. Previously, we were immediately exiting the decode loop and the
assert conditional that verified we were outputting frames correctly was
inverted, so all of our tests appeared to pass, even though the H264
vectors are actually failing due to a missing call to "drain()". This
issue is being tracked separately in b/390010500.

Note that because the C2Wrapper currently only supports I420 planar
frames, this also means adding some conversion functions to the V4L2
stateless queue. This change is intended to be temporary, and is
expecting to go away with the upcoming refactoring of cros-codecs's
frame management infrastructure.

Bug: 384980630
Test: test-25fps.vp8 and video.PlatformDecoding.cros_codecs*
Change-Id: I943b7e5c5bdcb85c3f6932de4e8e902c11abe86d