Rework queuing of v4l2 buffers
The allocation of buffers was based on returning early when
the decoder ran out of OUTPUT buffers. This required there to
be the same number of CAPTURE and OUTPUT buffers.

Because the hardware is not pipelined it can only use one buffer
at a time. While one buffer is decoding, the next can be filled.
An optimal decoder will only need 2 OUTPUT buffers and ideally
dpb CAPTURE buffers.

There is still a logic problem with the dpb and a magic number
is used to make the buffers flow through.

Bug: 382293969
Test: test-25fps.h264
Change-Id: I29e1fa9a8b87c25e50d561f68c0f8f26256c8e95