Attempt to dequeue capture buffers
Buffers only need to be dequeued when they are displayed. Once
they are displayed they do not need to be kept around, unless
they are references.

At the same time there is no reason to leave them queued after
they are done processing.

If the buffers are used for reference they only need to last
until the referencing buffers are decoded, not displayed.

Dequeuing them early sets up for freeing them up earlier.

Bug: 390647955
Test: test-25fps.h264
Test: test-25fps.vp8
Change-Id: Ia64842929c2a640d08bf6e27c217a2850257611a